+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Fixes
 - Fixed `with_id` deprecation message ([#1002](https://github.com/neptune-ai/neptune-client/pull/1002))
+- Fix passing None as deprecated parameter to deprecated_parameter decorator ([#1001](https://github.com/neptune-ai/neptune-client/pull/1001))
 
 ## neptune-client 0.16.7